backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
PyMySQL==1.1.1
cryptography==44.0.2
httpx==0.28.1
python-jose==3.3.0
passlib==1.7.4
pydantic[email]==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.1.0
